+++ src/progress-unix/progress_unix.ml
@@ -11,7 +11,7 @@ let stderr_if_tty =
   if Unix.(isatty stderr) then Format.err_formatter
   else Format.make_formatter (fun _ _ _ -> ()) (fun () -> ())
 
-let with_display x = Progress.with_display ~ppf:stderr_if_tty x
+let with_reporters x = Progress.with_reporters ~ppf:stderr_if_tty x
 let start x = Progress.start ~ppf:stderr_if_tty x
 let finalise = Progress.finalise
 let ( / ) = Progress.( / )